[phenyl-mongodb] Rename internal files
